requirements.txt
Flask
Flask-PyMongo
Werkzeug
Flask-Login
pymongo
python-dotenv
flask-bcrypt
flask-session
